remove old worker
