+++ src/site/twitch.tv/ChatAPI.ts
@@ -45,7 +45,6 @@ const data = reactive({
 });
 
 let flushTimeout: number | undefined;
-let live: boolean;
 
 export function useChatAPI(scroller?: Ref<InstanceType<typeof UiScrollableVue> | undefined>, bounds?: Ref<DOMRect>) {
 	const store = useStore();
@@ -62,30 +61,6 @@ export function useChatAPI(scroller?: Ref<InstanceType<typeof UiScrollableVue> |
 			data.userInput++;
 		}
 	});
-
-	if (!live) {
-		store.awaitWorkerNotify("entitlements:create", (v) => {
-			if (!v || !(v.cid && v.ref_id)) return true;
-
-			if (!data.entitledUsers[v.cid]) {
-				data.entitledUsers[v.cid] = {
-					BADGE: [],
-					PAINT: [],
-					EMOTE_SET: [],
-				};
-			}
-
-			if (v.slot === "BADGE" || v.slot === "PAINT") {
-				// clear the slot if it's badge or paint as there can only be one
-				data.entitledUsers[v.cid][v.slot as SevenTV.EntitlementKind] = [];
-			}
-
-			data.entitledUsers[v.cid][v.slot as SevenTV.EntitlementKind].push(v.ref_id);
-
-			return true;
-		});
-	}
-	live = true;
 
 	function addMessage(message: Twitch.ChatMessage): void {
 		if (data.paused) {
